v1.0.2.8
- Changed Region variables from int to float
- Updated Character for fractional speed movement
- Fixed Job not getting next task to start
- Added Open and Close methods to Menu
- Fixed Menu.Clear() disposing textures

+++ OP_Engine/Characters/Character.cs
@@ -288,8 +288,8 @@ namespace OP_Engine.Characters
         {
             if (Region != null)
             {
-                Region.X = (int)location.X;
-                Region.Y = (int)location.Y;
+                Region.X = location.X;
+                Region.Y = location.Y;
                 Travelled += Speed;
             }
         }
